requirements.txt
python-telegram-bot[job-queue]
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
psycopg2-binary
